Refactor doFetch function to remove console.log statements in doFetch.mjs
